+++ src/main/java/com/unstoppabledomains/resolution/DomainResolution.java
@@ -1,6 +1,7 @@
 package com.unstoppabledomains.resolution;
 
 import java.util.List;
+import java.util.Map;
 
 import com.unstoppabledomains.config.network.model.Network;
 import com.unstoppabledomains.exceptions.dns.DnsException;
@@ -97,6 +98,15 @@ public interface DomainResolution {
      */
     String getOwner(String domain) throws NamingServiceException;
     
+    /**
+     * Resolves owner addresses for many domains
+     * @param domains string list of domain names you want to resolve
+     * @return map of domain to owner addreses.
+     * @throws NamingServiceException
+     */
+    Map<String, String> getBatchOwners(List<String> domains) throws NamingServiceException;
+
+
     /**
      * Resolves dns records from a domain
      *