+++ src/main/java/com/unstoppabledomains/resolution/naming/service/ENS.java
@@ -16,8 +16,11 @@ import com.unstoppabledomains.resolution.dns.DnsRecord;
 import com.unstoppabledomains.resolution.dns.DnsRecordsType;
 import com.unstoppabledomains.util.Utilities;
 
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 public class ENS extends BaseNamingService {
 
@@ -73,6 +76,11 @@ public class ENS extends BaseNamingService {
       throw new NamingServiceException(NSExceptionCode.UnregisteredDomain, new NSExceptionParams("d", domain));
     }
     return owner;
+  }
+
+  @Override
+  public Map<String, String> batchOwners(List<String> domains) throws NamingServiceException {
+    throw new NamingServiceException(NSExceptionCode.NotImplemented, new NSExceptionParams("m|n", "batchOwners", getType().toString()));
   }
 
   @Override