+++ src/main/java/com/unstoppabledomains/resolution/naming/service/UNS.java
@@ -1,7 +1,6 @@
 
 package com.unstoppabledomains.resolution.naming.service;
 
-import com.unstoppabledomains.config.network.NetworkConfigLoader;
 import com.unstoppabledomains.exceptions.ContractCallException;
 import com.unstoppabledomains.exceptions.dns.DnsException;
 import com.unstoppabledomains.exceptions.ns.NSExceptionCode;
@@ -21,8 +20,10 @@ import java.math.BigInteger;
 import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.function.BiConsumer;
 
 public class UNS extends BaseNamingService {
   private final ProxyReader proxyReaderContract;
@@ -66,6 +67,7 @@ public class UNS extends BaseNamingService {
     return result;
   }
 
+  @Override
   public  String getOwner(String domain) throws NamingServiceException {
     try {
       BigInteger tokenID = getTokenID(domain);
@@ -82,12 +84,34 @@ public class UNS extends BaseNamingService {
   }
 
   @Override
+  public Map<String, String> batchOwners(List<String> domains) throws NamingServiceException {
+    Map<String, String> domainOwnerMap = new HashMap<>(domains.size());
+    try {
+      List<BigInteger> tokenIDs = new ArrayList<>();
+      for (String domain: domains) {
+        tokenIDs.add(getTokenID(domain));
+      }
+      
+      List<String> rawOwners = proxyReaderContract.batchOwners(tokenIDs.toArray(new BigInteger[tokenIDs.size()]));
+      Utilities.iterateSimultaneously(domains, rawOwners, (domain, rawOwner) -> {
+        String owner = Utilities.isEmptyResponse(rawOwner) ? null : rawOwner;
+        domainOwnerMap.put(domain, owner);
+      });
+      
+      return domainOwnerMap;
+    } catch(Exception e) {
+      throw configureNamingServiceException(e,
+        new NSExceptionParams("d|n", String.join(", ", domains),"UNS"));
+    }
+  }
+
+  @Override
   public List<DnsRecord> getDns(String domain, List<DnsRecordsType> types) throws NamingServiceException, DnsException {
     DnsUtils util = new DnsUtils();
     List<String> keys = constructDnsRecords(types);
     ProxyData data = resolveKeys(keys.toArray(new String[keys.size()]), domain);
     List<String> values = data.getValues();
-    Map<String, String> rawData = new HashMap();
+    Map<String, String> rawData = new HashMap<>();
     for (int i = 0; i < values.size(); i++) {
       rawData.put(keys.get(i), values.get(i));
     }
@@ -151,7 +175,7 @@ public class UNS extends BaseNamingService {
   }
 
   private List<String> constructDnsRecords(List<DnsRecordsType> types) {
-    List<String> records = new ArrayList();
+    List<String> records = new ArrayList<>();
     records.add("dns.ttl");
     for (DnsRecordsType type: types) {
       records.add("dns." + type.toString());
@@ -191,7 +215,7 @@ public class UNS extends BaseNamingService {
   }
 
 
-  private String owner(BigInteger tokenID) throws NamingServiceException {
+  private String owner(BigInteger tokenID) {
     return proxyReaderContract.getOwner(tokenID);
   }