+++ src/main/java/com/unstoppabledomains/util/Utilities.java
@@ -1,6 +1,8 @@
 package com.unstoppabledomains.util;
 
 import java.math.BigInteger;
+import java.util.Iterator;
+import java.util.function.BiConsumer;
 
 import com.unstoppabledomains.resolution.artifacts.Numeric;
 
@@ -50,5 +52,13 @@ public class Utilities {
    */
   public static BigInteger namehashToTokenID(String namehash) {
     return new BigInteger(namehash.substring(2), 16);
+  }
+
+  public static <T1, T2> void iterateSimultaneously(Iterable<T1> c1, Iterable<T2> c2, BiConsumer<T1, T2> consumer) {
+    Iterator<T1> i1 = c1.iterator();
+    Iterator<T2> i2 = c2.iterator();
+    while (i1.hasNext() && i2.hasNext()) {
+        consumer.accept(i1.next(), i2.next());
+    }
   }
 }